+++ design-system/heading/examples/00-basic.tsx
@@ -1,13 +1,15 @@
 import React from 'react';
 
-import { B400 } from '@atlaskit/theme/colors';
-import { token } from '@atlaskit/tokens';
+import {
+  UNSAFE_Box as Box,
+  UNSAFE_Stack as Stack,
+} from '@atlaskit/ds-explorations';
 
 import Heading from '../src';
 
 export default () => {
   return (
-    <div id="headings">
+    <Stack testId="headings" gap="scale.0">
       <Heading level="h100">h100</Heading>
       <Heading level="h200">h200</Heading>
       <Heading level="h300">h300</Heading>
@@ -17,15 +19,11 @@ export default () => {
       <Heading level="h700">h700</Heading>
       <Heading level="h800">h800</Heading>
       <Heading level="h900">h900</Heading>
-      <div
-        style={{
-          backgroundColor: token('color.background.brand.bold', B400),
-        }}
-      >
+      <Box backgroundColor="brand.bold">
         <Heading level="h900" color="inverse">
           inverse
         </Heading>
-      </div>
-    </div>
+      </Box>
+    </Stack>
   );
 };