+++ design-system/heading/src/index.tsx
@@ -1,3 +1,3 @@
 export { default } from './heading';
-
+export { default as HeadingContextProvider } from './heading-context';
 export type { HeadingProps } from './types';