+++ navigation/side-navigation/src/components/Section/__tests__/unit/index.spec.tsx
@@ -14,6 +14,9 @@ describe('Section', () => {
     onUnNest: jest.fn(),
     stack: [],
     parentId: ROOT_ID,
+    childIds: jest
+      .spyOn(React, 'useRef')
+      .mockReturnValue({ current: new Set<string>() }) as any,
   };
 
   it('should render a Section when contained in an active nested view', () => {