+++ design-system/onboarding/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/onboarding
 
+## 10.6.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.6.1
 
 ### Patch Changes