+++ design-system/section-message/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/section-message
 
+## 6.3.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 6.3.3
 
 ### Patch Changes